pyproject.toml
[project]
name = "practiceproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-groq>=0.0.2",
    "langchain-huggingface>=0.0.3",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.2",
]
